Add location notes support
